Step 2
Look at the data structure in the example.json Download example.json file, create a class Game that reflects this data structure. The class should be in its own file in a sub folder named models. The class should be exported as default and it should be imported in app.mjs. The class should be used throughout this project.