Step 3
Create a save function that saves a game to localStorage.
Create a function that retrieves all games that are saved in localStorage.
Create a function that outputs all the games as JSON
Create a function that can import the mentioned JSON and save all the games to localStorage
NB! You are not allowed to save games as an array, they must be saved as individual key/value items.