Step 4:
Add a input element with id “importSource” to the index.html
Inn app.mjs write the nessesary code to use a FileReader in conjunction with the code from step 3 so that the data from example.json can be imported to localStorage.
Also create an array games, that keeps an in memory record of all the games added.
The games array should always be populated when the application loads (app.mjs runs).

FileReader: https://developer.mozilla.org/en-US/docs/Web/API/FileReader

Had to go back to number three in order for the code to work.