Step 5:
Write the code that is necessary in app.mjs to add a visual record in index.html of each game (see gameRecordUI.png at the bottom of this page, for approximate goal Ticket To Ride used as an example). Note that there is a range input and a button that needs to be part of each games display. Do not add code for making these items interactive yet.
When done with this task each game should be listed out when viewing index.html.